+++ src/services/rbac/endpoint_to_permission_mapping.py
@@ -69,17 +69,6 @@ ENDPOINT_PERMISSION_MAPPING = {
         PATCH_METHOD: 'r8s:parent:update_parent',
         DELETE_METHOD: 'r8s:parent:remove_parent',
     },
-    '/parents/licenses/': {
-        GET_METHOD: 'r8s:parent:describe_parent',
-        POST_METHOD: 'r8s:parent:create_parent',
-        PATCH_METHOD: 'r8s:parent:update_parent',
-        DELETE_METHOD: 'r8s:parent:remove_parent',
-    },
-    '/parents/tenant-link/': {
-        GET_METHOD: 'r8s:parent:describe_tenant_link',
-        POST_METHOD: 'r8s:parent:link_tenant',
-        DELETE_METHOD: 'r8s:parent:unlink_tenant',
-    },
     '/parents/shape-rules/': {
         GET_METHOD: 'r8s:parent:describe_shape_rule',
         POST_METHOD: 'r8s:parent:create_shape_rule',
@@ -111,6 +100,9 @@ ENDPOINT_PERMISSION_MAPPING = {
     '/licenses/': {
         GET_METHOD: 'r8s:license:describe_license',
         DELETE_METHOD: 'r8s:license:delete_license',
+    },
+    '/licenses/sync/': {
+        POST_METHOD: 'r8s:license:sync_license'
     },
     '/settings/license-manager/config': {
         GET_METHOD: 'r8s:setting:describe_lm_config',